# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: FERNS — Federated Ecological Resource Network System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /gobotany:
    get:
      operationId: getGobotany
      tags: [gobotany]
      summary: Look up a Go Botany species page URL
      description: >
        Constructs and validates the Go Botany (Native Plant Trust) species page URL
        for a given scientific name. The URL is constructed from the genus and species
        epithet and validated with a live HTTP GET request (HTTP 200 = found). Results
        are cached 7 days. The species parameter must be a binomial name (genus + epithet);
        partial names return 400.
      parameters:
        - name: species
          in: query
          required: true
          description: >
            Binomial scientific name (e.g. Acer rubrum). Genus and species epithet required.
            Both components must contain only letters.
          schema:
            type: string
            minLength: 1
      responses:
        "200":
          description: URL lookup result (found or not-found — check the found field)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BotanicalWebRefResponse"
        "400":
          description: Missing or invalid species parameter
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /gobotany/species-text:
    get:
      operationId: getGobotanySpeciesText
      tags: [gobotany]
      summary: Scrape and return species page text from Go Botany
      description: >
        Fetches and caches botanical prose text from the Go Botany (Native Plant Trust)
        species detail page. The URL is constructed directly from the binomial name.
        Sections returned include Facts, Habitat, and Characteristics. First call
        scrapes live; subsequent calls return the cached text. Use ?refresh=true to
        force a re-scrape.
      parameters:
        - name: species
          in: query
          required: true
          description: Binomial scientific name (e.g. Acer rubrum)
          schema:
            type: string
            minLength: 1
        - name: refresh
          in: query
          required: false
          description: If "true", bypass cache and re-scrape the live page
          schema:
            type: string
            enum: ["true", "false"]
      responses:
        "200":
          description: Species page text result (found or not-found — check the found field)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SpeciesTextResponse"
        "400":
          description: Missing or malformed species parameter
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /illinois-wildflowers:
    get:
      operationId: getIllinoisWildflowers
      tags: [illinois-wildflowers]
      summary: Look up an Illinois Wildflowers species page URL
      description: >
        Looks up the Illinois Wildflowers (illinoiswildflowers.info) page URL for a
        given scientific name. Illinois Wildflowers organizes some species across
        multiple sections; multiple URLs may be returned in the data.results array.
        Species list is imported from the site and served from FERNS's local database.
        Results cached per database import cycle.
      parameters:
        - name: species
          in: query
          required: true
          description: Scientific name (e.g. Acer rubrum)
          schema:
            type: string
            minLength: 1
      responses:
        "200":
          description: URL lookup result (found or not-found — check the found field)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BotanicalWebRefResponse"
        "400":
          description: Missing species parameter
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /illinois-wildflowers/species-text:
    get:
      operationId: getIllinoisWildflowersSpeciesText
      tags: [illinois-wildflowers]
      summary: Scrape and return species page text from Illinois Wildflowers
      description: >
        Fetches and caches botanical prose text from the Illinois Wildflowers
        (illinoiswildflowers.info) species detail page. The page URL is resolved
        from the imported species list. Sections extracted include Description,
        Cultivation, Range & Habitat, and Comments. First call scrapes live;
        subsequent calls return the cached text. Use ?refresh=true to force a
        re-scrape. Returns found=false when the species is not in the imported list.
      parameters:
        - name: species
          in: query
          required: true
          description: Scientific name (e.g. Acer rubrum)
          schema:
            type: string
            minLength: 1
        - name: refresh
          in: query
          required: false
          description: If "true", bypass cache and re-scrape the live page
          schema:
            type: string
            enum: ["true", "false"]
      responses:
        "200":
          description: Species page text result (found or not-found — check the found field)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SpeciesTextResponse"
        "400":
          description: Missing species parameter
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
        message:
          type: string
    FernsProvenance:
      type: object
      description: >
        Provenance block present on every FERNS API response.
        Identity fields (source_id, fetched_at, method, upstream_url) are always present.
        Text fields (general_summary, technical_details) are conditionally present
        based on the provenance_verbosity query parameter (full|summary|none).
      required:
        - source_id
        - fetched_at
        - method
        - upstream_url
      properties:
        source_id:
          type: string
          description: Stable identifier for this data source (e.g. bonap-napa)
        fetched_at:
          type: string
          format: date-time
          description: When this record was obtained from the source
        method:
          type: string
          description: "How the data was obtained: api_fetch | blob_import | llm_synthesis"
        upstream_url:
          type: string
          description: Where this data came from (API endpoint, file path, or registry entry)
        general_summary:
          type: string
          description: Plain language description readable by a homeowner or community member
        technical_details:
          type: string
          description: >
            Research-grade description: methods, measurement protocols, algorithms,
            citations, and transformations — sufficient for a scientist to evaluate and reproduce
        matched_input:
          type: string
          description: >
            The normalized input that was actually used for this lookup (e.g., the name
            as queried). Present on endpoints that accept a name parameter.
...
    SpeciesTextResponse:
      type: object
      description: >
        Response shape for species page text scraping endpoints. Returned by
        /gobotany/species-text, /illinois-wildflowers/species-text,
        /minnesota-wildflowers/species-text, /missouri-plants/species-text,
        and /prairie-moon/species-text.
      required:
        - found
        - queried_at
        - cache_status
      properties:
        found:
          type: boolean
          description: True if the species page was found and text was extracted.
        queried_at:
          type: string
          format: date-time
        source_url:
          type: string
          description: The API endpoint URL that served this response.
        cache_status:
          type: string
          enum: [hit, miss, not_in_species_list]
          description: >
            "hit" — returned from cache. "miss" — live scrape performed and cached.
            "not_in_species_list" — species URL not in the imported species list;
            scrape skipped.
        fetch_error:
          type: string
          description: >
            Present only when a transient upstream error (network failure,
            timeout, 5xx) prevented the scrape. The result was NOT cached
            in this case and the next call will retry the live request.
        scraped_at:
          type: string
          format: date-time
          description: Timestamp of when the text was originally scraped.
        provenance:
          $ref: "#/components/schemas/FernsProvenance"
        data:
          nullable: true
          type: object
          description: Present when found=true. Null when not found.
          properties:
            species:
              type: string
              description: The species name as queried.
            url:
              type: string
              description: The page URL that was scraped.
            sections:
              type: object
              nullable: true
              additionalProperties:
                type: string
              description: >
                Labeled prose sections extracted from the page (e.g.
                Description, Cultivation, Facts, Habitat). Keys are section
                names, values are plain text content.
            full_text:
              type: string
              nullable: true
              description: >
                All sections concatenated as Label-colon-text blocks, separated by double newlines.
...
    BotanicalWebRefResponse:
      type: object
      description: >
        Standard response shape for botanical web reference source lookups
        (gobotany, google-images, illinois-wildflowers, minnesota-wildflowers,
        missouri-plants, prairie-moon, usda-plants).
      required:
        - found
        - queried_at
      properties:
        found:
          type: boolean
          description: >
            True if a direct species page URL was resolved. False for sources
            that cannot resolve a profile URL (usda-plants),
            which return a search_url instead.
        queried_at:
          type: string
          format: date-time
        source_url:
          type: string
        provenance:
          $ref: "#/components/schemas/FernsProvenance"
        data:
          nullable: true
          type: object
          description: Present when found=true (or when a search_url is returned). Null when not found.
          properties:
            species:
              type: string
              description: The species name as queried
            url:
              type: string
              nullable: true
              description: Direct species page URL. Null when only a search URL is available.
            search_url:
              type: string
              description: >
                Search URL when a direct profile URL cannot be constructed
                (usda-plants). Present instead of url for these sources.
            validation_method:
              type: string
              description: "How the URL was validated: http_get | species_list_lookup | direct_construction | not_resolvable"
            note:
              type: string
              description: Explanatory note when validation_method is not_resolvable
...